fix: coderabbit suggestion - don't pass empty list to query

+++ ee/backend/app/src/main/kotlin/io/tolgee/ee/service/qa/QaIssueService.kt
@@ -208,6 +208,10 @@ class QaIssueService(
       return
     }
 
+    if (types.isEmpty()) {
+      return
+    }
+
     qaIssueRepository.deleteAllByTranslationIdInAndTypeIn(translationIds, types)
   }
